Rename main.py to glycosylate.py and refactor structure handling
Renamed the primary script to better reflect its functionality. Moved complex structure and MTZ file handling logic to the interface module. Added a new script snfg.py for creating SNFG diagrams.
