+++ package/src/sails/__init__.py
@@ -1,23 +1,7 @@
-from .sails_module import (
-                           HKL,
-                           Reflection,
-                           Pair,
-                           Structure,
-                           Model,
-                           Chain,
-                           Residue,
-                           Atom,
-                           Position,
-                           Element,
-                           SeqId,
-                           n_glycosylate_from_objects,
-                           c_glycosylate_from_objects,
-                           Cell,
-                           MTZ,
-                           GlycoSite,
-                           Dot,
-                            test_snfg)
+from .sails_module import (HKL, Reflection, Pair, Structure, Model, Chain, Residue, Atom, Position, Element, SeqId,
+                           n_glycosylate_from_objects, c_glycosylate_from_objects, Cell, MTZ, GlycoSite, Dot, test_snfg,
+                           get_snfg, get_all_snfgs)
 from .__version__ import __version__
-from .main import glycosylate
-from .interface import extract_gemmi_mtz, extract_sails_mtz, extract_gemmi_structure, extract_sails_structure
-from .compare import compare
+from .glycosylate import glycosylate
+from .interface import extract_gemmi_mtz, extract_sails_mtz, extract_gemmi_structure, extract_sails_structure, \
+    get_sails_structure, get_sails_mtz
