+++ package/src/sails/interface.py
@@ -1,9 +1,88 @@
 import logging
 from typing import List, Tuple
-
+from pathlib import Path
 import gemmi
 import sails
 import numpy as np
+
+
+
+def read_sf_cif(mtz: Path):
+    """
+    :param mtz: Path to the CIF file containing structure factor data.
+    :return: MTZ file containing the converted structure factor data.
+
+    """
+    doc = gemmi.cif.read(str(mtz))
+    rblocks = gemmi.as_refln_blocks(doc)
+    if not rblocks:
+        raise RuntimeError("SF CIF supplied has no reflections")
+    cif2mtz = gemmi.CifToMtz()
+    return cif2mtz.convert_block_to_mtz(rblocks[0])
+
+
+def get_sails_mtz(mtz: gemmi.Mtz | Path | str, f: str, sigf: str, fwt: str, phwt: str):
+    """
+    :param mtz: Path to an MTZ file, an instance of gemmi.Mtz, or a string representing the path to an MTZ file.
+    :param f: Column name of the F values in the MTZ file.
+    :param sigf: Column name of the SIGF values in the MTZ file.
+    :param fwt: Column name of the FWT values in the MTZ file.
+    :param phwt: Column name of the PHWT values in the MTZ file.
+    :return: A gemmi.Mtz object containing only the specified F and SIGF columns.
+
+    This method extracts the specified F and SIGF columns from an MTZ file and returns a new MTZ object
+    containing only those columns. It supports three types of input for the 'mtz' parameter:
+    1) A string or Path object representing the path to the MTZ file.
+    2) An instance of the gemmi.Mtz class.
+    3) A gemmi.Mtz object.
+
+    If 'mtz' is an instance of the gemmi.Mtz class, the method directly extracts the specified columns
+    using the provided 'f' and 'sigf' column names.
+
+    If 'mtz' is a string or Path object representing the path to an MTZ file, the method first checks
+    the file extension to determine the format of the file. If the file has a CIF or ENT extension,
+    it is assumed to be a CIF file and the 'read_sf_cif' function is used to read the data. Otherwise,
+    the 'gemmi.read_mtz_file' function is used to read the MTZ file. The extracted columns are then
+    obtained from the resulting MTZ object using the specified 'f' and 'sigf' column names.
+
+    If 'mtz' is neither an instance of gemmi.Mtz, nor a string/Path object, a RuntimeError is raised.
+
+    The method returns a new gemmi.Mtz object containing only the specified F and SIGF columns.
+    """
+    if isinstance(mtz, gemmi.Mtz):
+        sails_mtz = extract_gemmi_mtz(mtz=mtz, column_names=[f, sigf, fwt, phwt])
+    elif isinstance(mtz, Path) or isinstance(mtz, str):
+        mtz = Path(mtz)
+        if ".cif" in mtz.suffixes or ".ent" in mtz.suffixes:
+            m = read_sf_cif(mtz)
+        else:
+            m = gemmi.read_mtz_file(str(mtz))
+        sails_mtz = extract_gemmi_mtz(mtz=m, column_names=[f, sigf, fwt, phwt])
+    else:
+        raise RuntimeError("Unknown object passed to second argument of n_glycosylate function, allowed types are"
+                           "gemmi.Mtz, Path, and str")
+    return sails_mtz
+
+
+def get_sails_structure(structure):
+    """
+    Retrieves the Sails structure from the provided input.
+
+    :param structure: The input structure. It can be either of type gemmi.Structure, or a file path as type Path
+                      or str.
+    :return: The Sails structure object.
+    :raises RuntimeError: If an unknown object is passed as the first argument. Only gemmi.Structure, Path, and str
+                          are allowed types.
+    """
+    if isinstance(structure, gemmi.Structure):
+        sails_structure = extract_gemmi_structure(structure=structure)
+    elif isinstance(structure, Path) or isinstance(structure, str):
+        s = gemmi.read_structure(str(structure))
+        sails_structure = extract_gemmi_structure(structure=s)
+    else:
+        raise RuntimeError("Unknown object passed to first argument of n_glycosylate function, allowed types are"
+                           "gemmi.Structure, Path, and str")
+    return sails_structure
 
 
 def extract_gemmi_structure(structure: gemmi.Structure):
